\name{dwcodebook-package}
\alias{dwcodebook-package}
\alias{dwcodebook}
\docType{package}
\title{
Create a simple DataWeb like codebook
}
\description{
This package aims to create a simple DataWeb like codebook from dataframes. The codebook will contain histograms for variables with more than 6 values and tables for variables that have less. In the future, tools will be added to simply the creation of the variable info dataframe.
}
\details{
\tabular{ll}{
Package: \tab dwcodebook\cr
Type: \tab Package\cr
Version: \tab 0.91\cr
Date: \tab 2012-12-19\cr
License: \tab Artistic 2.0\cr
}

createcb() creates the codebook and compiles it using pdflatex and mkindex. testcb() will show an example

}
\author{
Author: BJ Harshfield, Vince Carey, Ariana Huebner
Maintainer: BJ Harshfield <rebjh@channing.harvard.edu>

}

